mark stateless properties with annotations
