+++ compiler/backend/src/org/jetbrains/kotlin/codegen/ExpressionCodegen.java
@@ -2574,10 +2574,11 @@ public class ExpressionCodegen extends KtVisitor<StackValue, StackValue> impleme
             fieldName = KotlinTypeMapper.mapDefaultFieldName(propertyDescriptor, isDelegatedProperty);
         }
 
+        boolean isStatic = isStaticBackingField || (isDelegatedProperty && KotlinBuiltIns.isStatelessProperty(delegateType));
         return StackValue.property(propertyDescriptor, backingFieldOwner,
                                    typeMapper.mapType(
                                            isDelegatedProperty && forceField ? delegateType : propertyDescriptor.getOriginal().getType()),
-                                   isStaticBackingField, fieldName, callableGetter, callableSetter, receiver, this, resolvedCall);
+                                   isStatic, fieldName, callableGetter, callableSetter, receiver, this, resolvedCall);
     }
 
     @NotNull
