+++ compiler/backend/src/org/jetbrains/kotlin/codegen/ImplementationBodyCodegen.java
@@ -342,6 +342,8 @@ public class ImplementationBodyCodegen extends ClassBodyCodegen {
     protected void generateSyntheticParts() {
         generatePropertyMetadataArrayFieldIfNeeded(classAsmType);
 
+        generateStatelessPropertyInitializers();
+
         generateFieldForSingleton();
 
         generateCompanionObjectBackingFieldCopies();
