+++ libraries/tools/binary-compatibility-validator/reference-public-api/kotlin-stdlib.txt
@@ -2013,6 +2013,9 @@ public abstract interface class kotlin/properties/ReadWriteProperty {
 	public abstract fun setValue (Ljava/lang/Object;Lkotlin/reflect/KProperty;Ljava/lang/Object;)V
 }
 
+public abstract interface annotation class kotlin/properties/StatelessProperty : java/lang/annotation/Annotation {
+}
+
 public abstract interface class kotlin/ranges/ClosedFloatingPointRange : kotlin/ranges/ClosedRange {
 	public abstract fun contains (Ljava/lang/Comparable;)Z
 	public abstract fun isEmpty ()Z
